+++ src/boj/p24k/p24263.rs
@@ -3,7 +3,7 @@ use std::io::{BufRead, Write};
 
 #[allow(dead_code)]
 fn solve24262(reader: &mut impl BufRead, writer: &mut impl Write) {
-    let n = read_line(reader).parse::<usize>().unwrap();
+    let n = read_line(reader).parse::<i64>().unwrap();
 
     writeln!(writer, "{}", n).unwrap();
     writeln!(writer, "1").unwrap();
