+++ src/boj/p9k/p9063.rs
@@ -3,7 +3,7 @@ use std::io::{BufRead, Write};
 
 #[allow(dead_code)]
 fn solve9063(reader: &mut impl BufRead, writer: &mut impl Write) {
-    let n = read_line(reader).parse::<usize>().unwrap();
+    let n = read_line(reader).parse::<i64>().unwrap();
 
     let mut min_x = i32::MAX;
     let mut max_x = i32::MIN;
